% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Blob-class}
\alias{Blob}
\alias{Blob-class}
\title{Blob}
\description{
Blah blah blah
}
\section{Slots}{

\describe{
\item{\code{type}}{a character vector}
}}
\section{Methods}{
\itemize{
\item \code{draw(type = "small")}: Something about blobs
}}
\examples{
Blob(type = 'small')
}

