% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.r
\name{md}
\alias{md}
\title{Testing \code{Sexpr} with \code{roxygen2} (with \code{@md})}
\description{
In this example, a part of the documentation chunk is not parsed correctly,
namely, the \code{Sexpr} bit.
}
\details{
This doesn't work \Sexpr\link{results=text}{1+1}

But this does! \Sexpr{1+1}
}
