% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.r
\name{noMd-end}
\alias{noMd-end}
\title{Testing `Sexpr` with `roxygen2` (with `@noMd` at the end)}
\description{
In this example, a part of the codechunk IS not parsed correctly,
namely, the `Sexpr` bit, as we set the `@noMd` at the end.
}
\details{
\Sexpr[results=text]{1+1}
}
