% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.r
\name{noMd-middle}
\alias{noMd-middle}
\title{Testing `Sexpr` with `roxygen2` (with `@noMd` at the middle)}
\description{
In this example, a part of the codechunk IS not parsed correctly,
namely, the `Sexpr` bit, as we set the `@noMd` in the middle.
}
\details{
Actually, right after this text, and when calling `roxygenize` I see the warning

\preformatted{
Warning: @noMd [code.r#34]: has no parameters
}
}
