\docType{class}
\name{MyClass-class}
\alias{+,MyClass,numeric-method}
\alias{MyClass-class}
\title{Class \code{"MyClass"}}
\usage{
\S4method{+}{MyClass,numeric}(e1, e2)
}
\description{
A class to illustrate roxygen bugs
}
\section{Slots}{

\describe{
\item{\code{slot1}}{a numeric vector}

\item{\code{slot2}}{a numeric vector}
}}

